+++ myqueue/complete.py
@@ -75,8 +75,9 @@ commands = {
          '--traceback', '-A', '--all'],
     'workflow':
         ['-f', '--force', '--max-tasks', '-t', '--targets', '-p',
-         '--pattern', '-z', '--dry-run', '-v', '--verbose', '-q',
-         '--quiet', '-T', '--traceback']}
+         '--pattern', '-a', '--arguments', '-z', '--dry-run',
+         '-v', '--verbose', '-q', '--quiet', '-T',
+         '--traceback']}
 # End of computer generated data
 
 aliases = {'rm': 'remove',